feat: Add mode parameter to get_sensor_map function
The function get_sensor_map_request was modified to take a new parameter 'mode' which specifies a statistic function to be applied on the sensor values. The parameter 'mode' was added to the function get_sensor_map to propagate this parameter in the request to the server. By default, mode='AVERAGE'.